\name{flubase-package}
\alias{flubase-package}
\docType{package}
\title{Estimates a baseline free of influenza epidemic effects}
\description{
 This package contains a set of functions to estimate the mortality (or other indicator) baseline  
 free of influenza epidemics for one or more than one time series (age groups, gender, regions, etc) 
 using one of four methods made available, implemented respectively in \code{\link{baseRM}}, \code{\link{baseSA}},\code{\link{baseIt_RM}} 
 and \code{\link{baseIt_SA}}.
}
\details{
\tabular{ll}{
Package: \tab flubase\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-02-23\cr
License: \tab GPL >=2 \cr
}

}
\author{Nunes B, Natario I and Carvalho L.

Nunes B  baltazar.nunes@insa.min-saude.pt

}
\references{
 Nunes B, Natario I, Carvalho L. Time series methods for obtaining excess mortality. Submitted to Statistical Methods in Medical Research (2009).
 
 Serfling RE Methods for Current Statistical Analysis of Excess Mortality 
 Pneumonia-Influenza Deaths Public Heath Reports 1963; 78 6:494 506.
 
 K. Choi and S.B. Thacker An evaluation of influenza mortality surveillance
 1962-1979. American Journal of Epidemiology 1981; 113 3: 215 216.
 
 Lui K-J and Kendal A.P. Impact of influenza epidemics on mortality in
 the united states from October 1972 to May 1985. American Journal of
 Public Health 1987; 77(6):712 716.
 
 L. Simonsen, M.J. Clarke, D. Williamson, D.F. Stroup, N.H. Arden and L.B. Schonberger 
 The impact of influenza vaccination on seasonal mortality in the US elderly population. 
 American Journal of Public Health 1997; 87(12):1994 1950. 
}

\keyword{ package }
\seealso{\code{\link{flubase}}, \code{\link{baseRM}}, \code{\link{baseSA}}, \code{\link{baseIt_RM}}, \code{\link{baseIt_SA}}}

